simplifying multi-objective code
