+++ highway_env/envs/__init__.py
@@ -1,5 +1,4 @@
 from highway_env.envs.highway_env import *
-from highway_env.envs.mo_highway_env import *
 from highway_env.envs.merge_env import *
 from highway_env.envs.parking_env import *
 from highway_env.envs.summon_env import *
